# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Invite Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds/{guildId}/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [invites]
      summary: Get invite leaderboard for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
  /guilds/{guildId}/stats:
    get:
      operationId: getGuildStats
      tags: [invites]
      summary: Get overall guild statistics
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildStats"
...
  /guilds/{guildId}/members:
    get:
      operationId: getGuildMembers
      tags: [invites]
      summary: Get recent member join history
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Member join list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembersResponse"
  /guilds/{guildId}/invites/user/{userId}:
    get:
      operationId: getUserInviteStats
      tags: [invites]
      summary: Get invite stats for a specific user in a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User invite stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInviteStats"
  /guilds/{guildId}/invites/bonus:
    post:
      operationId: adjustBonusInvites
      tags: [invites]
      summary: Add or remove bonus invites for a user (admin)
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBonusBody"
      responses:
        "200":
          description: Updated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInviteStats"
  /guilds/{guildId}/invites/reset:
    post:
      operationId: resetUserInvites
      tags: [invites]
      summary: Reset a user's invites (admin)
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetInvitesBody"
      responses:
        "200":
          description: Invites reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: string
        username:
          type: string
        avatar:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        regular:
          type: integer
        bonus:
          type: integer
        fake:
          type: integer
        left:
          type: integer
        total:
          type: integer
      required:
        - rank
        - userId
        - username
        - regular
        - bonus
        - fake
        - left
        - total
    LeaderboardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - entries
        - total
        - page
        - limit
    GuildStats:
      type: object
      properties:
        totalJoins:
          type: integer
        totalLeaves:
          type: integer
        totalFake:
          type: integer
        netGrowth:
          type: integer
        topInviter:
          type: ["string", "null"]
        topInviterCount:
          type: ["integer", "null"]
        joinsTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - totalJoins
        - totalLeaves
        - totalFake
        - netGrowth
        - joinsTrend
        - recentActivity
    TrendPoint:
      type: object
      properties:
        date:
          type: string
        joins:
          type: integer
        leaves:
          type: integer
      required:
        - date
        - joins
        - leaves
    ActivityItem:
      type: object
      properties:
        type:
          type: string
          enum: [join, leave]
        userId:
          type: string
        username:
          type: string
        inviterId:
          type: ["string", "null"]
        inviterName:
          type: ["string", "null"]
        timestamp:
          type: string
        isFake:
          type: boolean
      required:
        - type
        - userId
        - username
        - timestamp
        - isFake
    MemberJoin:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        avatar:
          type: ["string", "null"]
        inviterId:
          type: ["string", "null"]
        inviterName:
          type: ["string", "null"]
        inviteCode:
          type: ["string", "null"]
        joinedAt:
          type: string
        leftAt:
          type: ["string", "null"]
        isFake:
          type: boolean
        accountAgeDays:
          type: ["integer", "null"]
      required:
        - id
        - userId
        - username
        - joinedAt
        - isFake
    MembersResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberJoin"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - members
        - total
        - page
        - limit
    UserInviteStats:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        avatar:
          type: ["string", "null"]
        regular:
          type: integer
        bonus:
          type: integer
        fake:
          type: integer
        left:
          type: integer
        total:
          type: integer
        rank:
          type: ["integer", "null"]
        invitedMembers:
          type: array
          items:
            $ref: "#/components/schemas/MemberJoin"
      required:
        - userId
        - username
        - regular
        - bonus
        - fake
        - left
        - total
    AdjustBonusBody:
      type: object
      properties:
        userId:
          type: string
        amount:
          type: integer
        reason:
          type: ["string", "null"]
      required:
        - userId
        - amount
    ResetInvitesBody:
      type: object
      properties:
        userId:
          type: string
        type:
          type: string
          enum: [all, regular, bonus, fake, left]
      required:
        - userId
        - type
...